Upgrade Parallel
Okay this one was weird, the reference to the Parallel gem in `pet.rb` just, stopped working? Is that some weird downstream consequence of something we changed today, or has it just been broken for a long time and we just never ran that codepath? Seems… odd if we hadn't? But ok?

In any case, upgrading the gem seemed to fix whatever was causing it to not load in for whatever reason. Ok!